update carriage return fix to be backwards compatible
The use of carriage return fixes the apparent stripping of new lines on
rxvt-unicode 9.15, but when pasting into vim in 9.14 (on ubuntu at
least), vim will honor both the new line and the carriage return
resulting in extra lines between the pasted text. Resolve this by
replacing the trailing new lines with the carriage returns.